+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 4.0.2
 
 - fixed bug which caused `TBSMParallelState` not to setup all sub machines when performing transition into one state machine
+- fixed a bug which caused events not to bubble up to super states when not being handled by a sub state
+- **TODO** reworked event handling and state switching
 
 ## 4.0.1